Corrigido arquivos que não subiram corretamente
