+++ dash/admin.py
@@ -1,3 +1,6 @@
 from django.contrib import admin
+from .models import DadosSaresp
+
+admin.site.register(DadosSaresp)
 
 # Register your models here.
