+++ dash/templates/escola.html
@@ -1,77 +1,40 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>UNIVESP - PI</title>
-  <!-- Inclua os arquivos CSS do Bootstrap -->
-  <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.5.0/css/bootstrap.min.css">
-</head>
-<body>
+{% extends 'escolas.html' %}
+{% block charts %}
+  const ctx = document.getElementById('myChart');
+  const labels = ['% Acer. Port.','% Acer. Mat.', '% Acer. Cie.'];
 
-  <nav class="navbar navbar-expand-lg navbar-dark bg-dark">
-    <a class="navbar-brand" href="#">Estatísticas</a>
-    <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarNav" aria-controls="navbarNav" aria-expanded="false" aria-label="Toggle navigation">
-      <span class="navbar-toggler-icon"></span>
-    </button>
-    <div class="collapse navbar-collapse" id="navbarNav">
-      <ul class="navbar-nav">
-        <li class="nav-item active">
-          <a class="nav-link" href="#">Início</a>
-        </li>
-        <li class="nav-item">
-          <a class="nav-link" href="#">Dados</a>
-        </li>
-        <li class="nav-item">
-          <a class="nav-link" href="#">Análises</a>
-        </li>
-        <li class="nav-item">
-          <a class="nav-link" href="#">Gráficos</a>
-        </li>
-        <li class="nav-item">
-          <a class="nav-link" href="#">Sobre</a>
-        </li>
-      </ul>
-    </div>
-  </nav>
+  const data = {
+    labels: labels,
+    datasets: [
+        {% for serie in escola %}
+            {
+                label: '{{ serie.serie_ano }}',
+                data: [{{ serie.porc_cert_lp }}, {{ serie.porc_cert_mat }}, {{ serie.porc_cert_cie }}],
+            },
+        {% endfor %}
+    ]
+    };
 
-  <div class="container mt-5">
-    <h1>Bem-vindo ao site analítico dos dados SARESP 2022!</h1>
-    <p>Aqui você encontrará dados, análises e gráficos estatísticos.</p>
-    <div class="row mt-5">
-      <div class="col-md-4">
-        <div class="card">
-          <div class="card-body">
-            <h5 class="card-title">Dados</h5>
-            <p class="card-text">Explore nossa coleção de dados estatísticos.</p>
-            <a href="#" class="btn btn-primary">Ver Dados</a>
-          </div>
-        </div>
-      </div>
-      <div class="col-md-4">
-        <div class="card">
-          <div class="card-body">
-            <h5 class="card-title">Análises</h5>
-            <p class="card-text">Obtenha insights valiosos com nossas análises estatísticas.</p>
-            <a href="#" class="btn btn-primary">Ver Análises</a>
-          </div>
-        </div>
-      </div>
-      <div class="col-md-4">
-        <div class="card">
-          <div class="card-body">
-            <h5 class="card-title">Gráficos</h5>
-            <p class="card-text">Visualize os dados em gráficos interativos.</p>
-            <a href="#" class="btn btn-primary">Ver Gráficos</a>
-          </div>
-        </div>
-      </div>
-    </div>
-  </div>
+  new Chart(ctx, {
+    type: 'line',
+    data: data,
+    options: {
+        responsive: true,
+        plugins: {
+          legend: {
+            position: 'top',
+          },
+          title: {
+            display: true,
+            text: 'Dados da Escola: '
+          }
+        }
+    },
+  });
 
-  <!-- Inclua os arquivos JavaScript do Bootstrap -->
-  <script src="https://code.jquery.com/jquery-3.5.1.slim.min.js"></script>
-  <script src="https://cdn.jsdelivr.net/npm/@popperjs/core@2.5.3/dist/umd/popper.min.js"></script>
-  <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.5.0/js/bootstrap.min.js"></script>
-</body>
-</html>
+{% endblock %}
+{% block dados_escola %}
+        <div>
+            <canvas id="myChart"></canvas>
+        </div>
+{% endblock %}
