+++ dash/templates/escolas.html
@@ -1,4 +1,11 @@
 {% extends 'base.html' %}
 {% block content %}
-    <h1>Escola "{{ escola }}"!</h1>
+    <select onchange="escola()" id="escolas">
+        <option disabled SELECTED>Selecione uma escola</option>
+        {% for escola in escolas %}
+            <option value="{{ escola.codesc }}">{{ escola.escola }}</option>
+        {% endfor %}
+    </select>
+    {% block dados_escola %}
+    {% endblock %}
 {% endblock %}
