+++ dash/templates/index.html
@@ -1,0 +1,34 @@
+{% extends 'base.html' %}
+{% block content %}
+<h1>Bem-vindo ao site analítico dos dados SARESP 2022!</h1>
+    <p>Aqui você encontrará dados, análises e gráficos estatísticos.</p>
+    <div class="row mt-5">
+      <div class="col-md-4">
+        <div class="card">
+          <div class="card-body">
+            <h5 class="card-title">Dados</h5>
+            <p class="card-text">Explore nossa coleção de dados estatísticos.</p>
+            <a href="#" class="btn btn-primary">Ver Dados</a>
+          </div>
+        </div>
+      </div>
+      <div class="col-md-4">
+        <div class="card">
+          <div class="card-body">
+            <h5 class="card-title">Análises</h5>
+            <p class="card-text">Obtenha insights valiosos com nossas análises estatísticas.</p>
+            <a href="#" class="btn btn-primary">Ver Análises</a>
+          </div>
+        </div>
+      </div>
+      <div class="col-md-4">
+        <div class="card">
+          <div class="card-body">
+            <h5 class="card-title">Gráficos</h5>
+            <p class="card-text">Visualize os dados em gráficos interativos.</p>
+            <a href="#" class="btn btn-primary">Ver Gráficos</a>
+          </div>
+        </div>
+      </div>
+    </div>
+{% endblock %}
